+++ database/seeders/PostsMediaLocalSeeder.php
@@ -1,17 +1,19 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Helpers\CuriousPeople\CuriousNum;
 use App\Helpers\CuriousPeople\CuriousStorage;
+use App\Helpers\CuriousPeople\CuriousUrl;
 use App\Post;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Seeder;
-use Spatie\Tags\Tag;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Output\ConsoleOutput;
 
 //use Faker\Generator as Faker;
 
-class PostsTagsSeeder extends Seeder
+class PostsMediaLocalSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -27,46 +29,44 @@ class PostsTagsSeeder extends Seeder
         $faker = Faker\Factory::create();
         Model::unguard();
         $this->setFKCheckOff();
+        DB::table('media')->where('model_type', \App\Post::class)->delete();
 
         // POST
         $posts_count = Post::count();
         $output = new ConsoleOutput();
         $progress = new ProgressBar($output, $posts_count);
         $progress->setFormat('%message%: %current%/%max% [%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s% %memory:6s%'."\n");
-        $progress->setMessage('Generating Post Tags');
+        $progress->setMessage('Generating Post Images');
         $progress->start();
 
         Post::chunk(100, function ($posts) use ($progress, $posts_count, $faker) {
+            $output_model = new ConsoleOutput();
+            $progress_model = new ProgressBar($output_model, $posts_count);
+            $progress_model->setFormat('%message%: %current%/%max% [%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s% %memory:6s%'."\n");
+            $progress_model->setMessage('Creating Images');
+            $progress_model->start();
+
             foreach ($posts as $post) {
-                $output_tags = new ConsoleOutput();
-                $progress_tags = new ProgressBar($output_tags, $posts_count);
-                $progress_tags->setFormat('%message%: %current%/%max% [%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s% %memory:6s%'."\n");
-                $progress_tags->setMessage('Creating Tags');
-                $progress_tags->start();
+                // add remote images
+                $num = CuriousNum::getRandomBias(config('seeder.posts.media'));
 
-                $tags_count = CuriousNum::getRandomBias(config('seeder.posts.hashtags'));
-                for ($i = 0; $i < $tags_count; $i++) {
-                    $tag = $faker->word;
-                    $this_tag = Tag::findOrCreate($tag, 'hashtag');
-                    $post->attachTag($this_tag);
-                    $progress_tags->advance();
+                $addimg = rand(0, 10);
+                if ($post->type == 'image' || ($addimg > 7 && $post->type != 'text')) {
+                    $img = CuriousStorage::randomFileFromPath('/public/seeder/covers/');
+                    $url = storage_path('app/'.$img);
+                    $post->addMedia($url)
+                    ->usingFileName(Str::uuid())
+                    ->preservingOriginal()
+                    ->toMediaCollection('cover');
                 }
 
-                $cats_count = CuriousNum::getRandomBias(config('seeder.posts.categories'));
-                for ($i = 0; $i < $cats_count; $i++) {
-                    $tag = $faker->word;
-                    $this_tag = Tag::findOrCreate('bar', 'category');
-                    $post->attachTag($this_tag);
-                    $progress_tags->advance();
-                }
-
-                $progress_tags->clear();
-                $progress->advance();
+                $progress_model->advance();
             }
-            $progress_tags->clear();
-            $progress_tags->finish();
-            $progress->clear();
+            $progress_model->clear();
+            $progress_model->finish();
+            $progress->advance();
         });
+        $progress->clear();
         $progress->finish();
 
         $this->setFKCheckOn();
