+++ database/seeders/PostsTagsSeeder.php
@@ -1,8 +1,10 @@
 <?php
 
-use App\Board;
+namespace Database\Seeders;
+
 use App\Helpers\CuriousPeople\CuriousNum;
 use App\Helpers\CuriousPeople\CuriousStorage;
+use App\Post;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Seeder;
 use Spatie\Tags\Tag;
@@ -11,7 +13,7 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 
 //use Faker\Generator as Faker;
 
-class BoardsTagsSeeder extends Seeder
+class PostsTagsSeeder extends Seeder
 {
     /**
      * Run the database seeds.
@@ -28,35 +30,35 @@ class BoardsTagsSeeder extends Seeder
         Model::unguard();
         $this->setFKCheckOff();
 
-        // BOARD
-        $boards_count = Board::count();
+        // POST
+        $posts_count = Post::count();
         $output = new ConsoleOutput();
-        $progress = new ProgressBar($output, $boards_count);
+        $progress = new ProgressBar($output, $posts_count);
         $progress->setFormat('%message%: %current%/%max% [%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s% %memory:6s%'."\n");
-        $progress->setMessage('Generating Board Tags');
+        $progress->setMessage('Generating Post Tags');
         $progress->start();
 
-        Board::chunk(100, function ($boards) use ($progress, $boards_count, $faker) {
-            foreach ($boards as $board) {
+        Post::chunk(100, function ($posts) use ($progress, $posts_count, $faker) {
+            foreach ($posts as $post) {
                 $output_tags = new ConsoleOutput();
-                $progress_tags = new ProgressBar($output_tags, $boards_count);
+                $progress_tags = new ProgressBar($output_tags, $posts_count);
                 $progress_tags->setFormat('%message%: %current%/%max% [%bar%] %percent:3s%% %elapsed:6s%/%estimated:-6s% %memory:6s%'."\n");
                 $progress_tags->setMessage('Creating Tags');
                 $progress_tags->start();
 
-                $tags_count = CuriousNum::getRandomBias(config('seeder.boards.hashtags'));
+                $tags_count = CuriousNum::getRandomBias(config('seeder.posts.hashtags'));
                 for ($i = 0; $i < $tags_count; $i++) {
                     $tag = $faker->word;
                     $this_tag = Tag::findOrCreate($tag, 'hashtag');
-                    $board->attachTag($this_tag);
+                    $post->attachTag($this_tag);
                     $progress_tags->advance();
                 }
 
-                $cats_count = CuriousNum::getRandomBias(config('seeder.boards.categories'));
+                $cats_count = CuriousNum::getRandomBias(config('seeder.posts.categories'));
                 for ($i = 0; $i < $cats_count; $i++) {
                     $tag = $faker->word;
-                    $this_tag = Tag::findOrCreate($tag, 'category');
-                    $board->attachTag($this_tag);
+                    $this_tag = Tag::findOrCreate('bar', 'category');
+                    $post->attachTag($this_tag);
                     $progress_tags->advance();
                 }
